refactor docs site to use PageContent and friends
