+++ packages/visual-stack-docs/src/containers/Components/Docs/spinner.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Panel, Body, Header } from '@cjdev/visual-stack/lib/components/Panel';
-import { Demo, Snippet } from './Demo';
+import { Demo, Snippet } from '../../../components/Demo';
 
 import Spinner from '@cjdev/visual-stack/lib/components/Spinner';
