+++ packages/visual-stack-docs/src/containers/Icons/index.js
@@ -1,6 +1,16 @@
 import React from 'react';
 
+import PageContent from '@cjdev/visual-stack/lib/components/PageContent';
+import { PageHeader, PageTitle } from '@cjdev/visual-stack/lib/components/PageHeader';
+
 const Icons = () =>
-  <div>Icons</div>;
+  <div>
+    <PageHeader>
+      <PageTitle>Icons</PageTitle>
+    </PageHeader>
+    <PageContent>
+    </PageContent>
+  </div>;
+
 
 export default Icons;
